requirements.txt
fastapi
uvicorn
websockets
openai
TTS
openai-whisper
numpy
soundfile